pulling out non-class provider functionality
