+++ main/provider/index.ts
@@ -5,42 +5,29 @@ import log from 'electron-log'
 import utils from 'web3-utils'
 import { recoverTypedMessage, Version } from 'eth-sig-util'
 import crypto from 'crypto'
-
 import {
-  padToEven,
-  unpadHexString,
   addHexPrefix,
-  stripHexPrefix,
   intToHex,
   isHexString,
   isHexPrefixed,
   fromUtf8,
-  toBuffer,
-  pubToAddress,
-  ecrecover,
-  hashPersonalMessage,
 } from 'ethereumjs-util'
 
 import store from '../store'
-import protectedMethods from '../api/protectedMethods'
 import packageFile from '../../package.json'
 
 import proxyConnection from './proxy'
 import accounts, { AccountRequest, TransactionRequest, SignTypedDataRequest, AddChainRequest, AddTokenRequest } from '../accounts'
 import Chains, { Chain } from '../chains'
 import { getType as getSignerType, Type as SignerType } from '../signers/Signer'
-import { populate as populateTransaction, usesBaseFee, maxFee, TransactionData } from '../transaction'
+import { populate as populateTransaction, maxFee, TransactionData } from '../transaction'
 import FrameAccount from '../accounts/Account'
 import { capitalize, arraysMatch } from '../../resources/utils'
 import { ApprovalType } from '../../resources/constants'
-
-const NATIVE_CURRENCY = '0x0000000000000000000000000000000000000000'
-
-const permission = (date: number, method: string) => ({ parentCapability: method, date })
+import { ecRecover, feeTotalOverMax, gasFees, getActiveChains, getAssets, getChains, getPermissions, getRawTx, getSignedAddress, isCurrentAccount, isScanning, loadAssets, requestPermissions, resError } from './helpers'
 
 type Origins = Record<string, { chainId: number }>
 type Subscriptions = { [key in SubscriptionType]: string[] }
-type Balance = Token & { balance: string, displayBalance: string }
 
 interface RequiredApproval {
   type: ApprovalType,
@@ -50,51 +37,6 @@ interface RequiredApproval {
 export interface TransactionMetadata {
   tx: TransactionData,
   approvals: RequiredApproval[]
-}
-
-function isScanning (account: Address) {
-  const lastUpdated = store('main.accounts', account, 'balances.lastUpdated') as number
-  return !lastUpdated || (new Date().getTime() - lastUpdated) > (1000 * 60 * 5)
-}
-
-function getNativeCurrency (chainId: number) {
-  const currency = store('main.networksMeta.ethereum', chainId, 'nativeCurrency')
-
-  return (currency || { usd: { price: 0 } }) as Currency
-}
-
-function getRate (address: Address) {
-  const rate = store('main.rates', address.toLowerCase())
-
-  return (rate || { usd: { price: 0 } }) as Rate
-}
-
-function loadAssets (accountId: string) {
-  const balances: Balance[] = store('main.balances', accountId) || []
-
-  const response = { nativeCurrency: [] as RPC.GetAssets.NativeCurrency[], erc20: [] as RPC.GetAssets.Erc20[] }
-
-  return balances.reduce((assets, balance) => {
-    if (balance.address === NATIVE_CURRENCY) {
-      const currency = getNativeCurrency(balance.chainId)
-
-      assets.nativeCurrency.push({
-        ...balance,
-        currencyInfo: currency
-      })
-    } else {
-      const { usd } = getRate(balance.address)
-
-      assets.erc20.push({
-        ...balance,
-        tokenInfo: {
-          lastKnownPrice: { usd }
-        }
-      })
-    }
-
-    return assets
-  }, response)
 }
 
 export class Provider extends EventEmitter {
@@ -177,17 +119,6 @@ export class Provider extends EventEmitter {
     })
   }
 
-  getCoinbase (payload: RPCRequestPayload, res: RPCRequestCallback) {
-    accounts.getAccounts((err, accounts) => {
-      if (err) return this.resError(`signTransaction Error: ${JSON.stringify(err)}`, payload, res)
-      res({ id: payload.id, jsonrpc: payload.jsonrpc, result: (accounts || [])[0] })
-    })
-  }
-
-  getAccounts (payload: JSONRPCRequestPayload, res: RPCRequestCallback) {
-    res({ id: payload.id, jsonrpc: payload.jsonrpc, result: accounts.getSelectedAddresses().map(a => a.toLowerCase()) })
-  }
-
   getNetVersion (payload: RPCRequestPayload, res: RPCRequestCallback, targetChain: Chain) {
     const { type, id } = (targetChain || store('main.origins', uuidv5(payload._origin, uuidv5.DNS)))
 
@@ -214,10 +145,6 @@ export class Provider extends EventEmitter {
     res({ id: payload.id, jsonrpc: payload.jsonrpc, ...response })
   }
 
-  getChains (payload: JSONRPCRequestPayload, res: RPCSuccessCallback) {
-    res({ id: payload.id, jsonrpc: payload.jsonrpc, result: getActiveChains().map(intToHex) })
-  }
-
   declineRequest (req: AccountRequest) {
     const res = (data: any) => {
       if (this.handlers[req.handlerId]) this.handlers[req.handlerId](data)
@@ -225,37 +152,7 @@ export class Provider extends EventEmitter {
     }
 
     const payload = req.payload
-    this.resError({ message: 'User declined transaction', code: 4001 }, payload, res)
-  }
-
-  resError (errorData: string | EVMError, request: RPCId, res: RPCErrorCallback) {
-    const error = (typeof errorData === 'string')
-      ? { message: errorData, code: -1 }
-      : { message: errorData.message, code: errorData.code || -1 }
-
-    log.warn(error)
-    res({ id: request.id, jsonrpc: request.jsonrpc, error })
-  }
-
-  getSignedAddress (signed: string, message: string, cb: Callback<String>) {
-    const signature = Buffer.from(signed.replace('0x', ''), 'hex')
-    if (signature.length !== 65) cb(new Error('Frame verifySignature: Signature has incorrect length'))
-    let v = signature[64]
-    v = v === 0 || v === 1 ? v + 27 : v
-    const r = toBuffer(signature.slice(0, 32))
-    const s = toBuffer(signature.slice(32, 64))
-    const hash = hashPersonalMessage(toBuffer(message))
-    const verifiedAddress = '0x' + pubToAddress(ecrecover(hash, v, r, s)).toString('hex')
-    cb(null, verifiedAddress)
-  }
-
-  ecRecover (payload: JSONRPCRequestPayload, res: RPCRequestCallback) {
-    const message = payload.params[0]
-    const signed = payload.params[1]
-    this.getSignedAddress(signed, message, (err, verifiedAddress) => {
-      if (err) return this.resError(err.message, payload, res)
-      res({ id: payload.id, jsonrpc: payload.jsonrpc, result: verifiedAddress })
-    })
+    resError({ message: 'User declined transaction', code: 4001 }, payload, res)
   }
 
   verifySignature (signed: string, message: string, address: string, cb: Callback<boolean>) {
@@ -273,7 +170,7 @@ export class Provider extends EventEmitter {
       }
     }
 
-    this.getSignedAddress(signed, message, (err, verifiedAddress) => {
+    getSignedAddress(signed, message, (err, verifiedAddress) => {
       if (err) return cb(err)
       if ((verifiedAddress || '').toLowerCase() !== address.toLowerCase()) return cb(new Error('Frame verifySignature: Failed ecRecover check'))
       cb(null, true)
@@ -301,13 +198,13 @@ export class Provider extends EventEmitter {
 
     accounts.signMessage(address, message, (err, signed) => {
       if (err) {
-        this.resError(err.message, payload, res)
+        resError(err.message, payload, res)
         cb(err, undefined)
       } else {
         const signature = signed || ''
         this.verifySignature(signature, message, address, err => {
           if (err) {
-            this.resError(err.message, payload, res)
+            resError(err.message, payload, res)
             cb(err)
           } else {
             res({ id: payload.id, jsonrpc: payload.jsonrpc, result: signature })
@@ -331,7 +228,7 @@ export class Provider extends EventEmitter {
 
     accounts.signTypedData(req.version, address, dataToSign, (err, signature) => {
       if (err) {
-        this.resError(err.message, payload, res)
+        resError(err.message, payload, res)
         cb(err)
       } else {
         const sig = signature || ''
@@ -344,19 +241,12 @@ export class Provider extends EventEmitter {
           res({ id: payload.id, jsonrpc: payload.jsonrpc, result: sig })
           cb(null, sig)
         } catch (err: any) {
-          this.resError(err.message, payload, res)
+          resError(err.message, payload, res)
 
           cb(err)
         }
       }
     })
-  }
-
-  feeTotalOverMax (rawTx: TransactionData, maxTotalFee: number) {
-    const maxFeePerGas = usesBaseFee(rawTx) ? parseInt(rawTx.maxFeePerGas || '', 16) : parseInt(rawTx.gasPrice || '', 16)
-    const gasLimit = parseInt(rawTx.gasLimit || '', 16)
-    const totalFee = maxFeePerGas * gasLimit
-    return totalFee > maxTotalFee
   }
 
   signAndSend (req: TransactionRequest, cb: Callback<string>) {
@@ -369,7 +259,7 @@ export class Provider extends EventEmitter {
     const payload = req.payload
     const maxTotalFee = maxFee(rawTx)
 
-    if (this.feeTotalOverMax(rawTx, maxTotalFee)) {
+    if (feeTotalOverMax(rawTx, maxTotalFee)) {
       const chainId = parseInt(rawTx.chainId)
       const symbol = store(`main.networks.ethereum.${chainId}.symbol`)
       const displayAmount = symbol
@@ -378,12 +268,12 @@ export class Provider extends EventEmitter {
 
       const err = `Max fee is over hard limit${displayAmount}`
 
-      this.resError(err, payload, res)
+      resError(err, payload, res)
       cb(new Error(err))
     } else {
       accounts.signTransaction(rawTx, (err, signedTx) => { // Sign Transaction
         if (err) {
-          this.resError(err, payload, res)
+          resError(err, payload, res)
           cb(err)
         } else {
           accounts.setTxSigned(req.handlerId, err => {
@@ -400,7 +290,7 @@ export class Provider extends EventEmitter {
                 if (done) return
                 done = true
                 if (response.error) {
-                  this.resError(response.error, payload, res)
+                  resError(response.error, payload, res)
                   cb(new Error(response.error.message))
                 } else {
                   res(response)
@@ -445,21 +335,6 @@ export class Provider extends EventEmitter {
     })
   }
 
-  private getRawTx (newTx: RPC.SendTransaction.TxParams): TransactionData {
-    const { gas, gasLimit, gasPrice, data, value, type, ...rawTx } = newTx
-    const parsedValue = !value || parseInt(value, 16) === 0 ? '0x0' : addHexPrefix(unpadHexString(value) || '0')
-
-    return {
-      ...rawTx,
-      from: rawTx.from || ((accounts.current() || {}).id),
-      type: '0x0',
-      value: parsedValue,
-      data: addHexPrefix(padToEven(stripHexPrefix(data || '0x'))),
-      gasLimit: gasLimit || gas,
-      chainId: rawTx.chainId
-    }
-  }
-
   private addRequestHandler (res: RPCRequestCallback) {
     const handlerId: string = uuid()
     this.handlers[handlerId] = res
@@ -491,15 +366,6 @@ export class Provider extends EventEmitter {
         return resolve(addHexPrefix(paddedLimit.toString(16)))
       }, targetChain)
     })
-  }
-
-  private isCurrentAccount (address: string, account = accounts.current()) {
-    const accountToCheck = account || accounts.current() || { id: '' }
-    return address && (accountToCheck.id.toLowerCase() === address.toLowerCase())
-  }
-
-  private gasFees (rawTx: TransactionData) {
-    return store('main.networksMeta', 'ethereum', parseInt(rawTx.chainId, 16), 'gas')
   }
 
   getNonce (rawTx: TransactionData, res: RPCRequestCallback) {
@@ -556,8 +422,9 @@ export class Provider extends EventEmitter {
 
     try {
       const approvals: RequiredApproval[] = []
-      const rawTx = this.getRawTx(newTx)
-      const gas = this.gasFees(rawTx)
+      const accountId = (accounts.current() || {}).id
+      const rawTx = getRawTx(newTx, accountId)
+      const gas = gasFees(rawTx)
       const chainConfig = this.connection.connections['ethereum'][parseInt(rawTx.chainId)].chainConfig
 
       const estimateGas = rawTx.gasLimit
@@ -599,7 +466,7 @@ export class Provider extends EventEmitter {
     const txChain = txParams.chainId
 
     if (targetChain && txChain && targetChain !== txChain) {
-      return this.resError(`Chain for transaction (${txChain}) does not match request target chain (${targetChain})`, payload, res)
+      return resError(`Chain for transaction (${txChain}) does not match request target chain (${targetChain})`, payload, res)
     }
 
     const newTx = {
@@ -613,12 +480,12 @@ export class Provider extends EventEmitter {
 
     this.fillTransaction(newTx, (err, transactionMetadata) => {
       if (err) {
-        this.resError(err, payload, res)
+        resError(err, payload, res)
       } else {
         const txMetadata = transactionMetadata as TransactionMetadata
         const from = txMetadata.tx.from
 
-        if (from && !this.isCurrentAccount(from, currentAccount)) return this.resError('Transaction is not from currently selected account', payload, res)
+        if (from && !isCurrentAccount(from, currentAccount)) return resError('Transaction is not from currently selected account', payload, res)
 
         const handlerId = this.addRequestHandler(res)
         const { feesUpdated, ...data } = txMetadata.tx
@@ -676,7 +543,7 @@ export class Provider extends EventEmitter {
     const from = orderedParams[0]
     const currentAccount = accounts.current()
 
-    if (!this.isCurrentAccount(from, currentAccount)) return this.resError('sign request is not from currently selected account.', payload, res)
+    if (!isCurrentAccount(from, currentAccount)) return resError('sign request is not from currently selected account.', payload, res)
 
     const handlerId = this.addRequestHandler(res)
 
@@ -706,7 +573,7 @@ export class Provider extends EventEmitter {
     const targetAccount = accounts.get(from.toLowerCase())
 
     if (!targetAccount) {
-      return this.resError(`unknown account: ${from}`, payload, res)
+      return resError(`unknown account: ${from}`, payload, res)
     }
 
     // HACK: Standards clearly say, that second param is an object but it seems like in the wild it can be a JSON-string.
@@ -715,7 +582,7 @@ export class Provider extends EventEmitter {
         typedData = JSON.parse(typedData)
         payload.params = [from, typedData, ...additionalParams]
       } catch (e) {
-        return this.resError('Malformed typedData.', payload, res)
+        return resError('Malformed typedData.', payload, res)
       }
     }
 
@@ -724,7 +591,7 @@ export class Provider extends EventEmitter {
     // check for signers that only support signing a specific version of typed data
     if (version !== 'V4' && signerType && [SignerType.Ledger, SignerType.Lattice, SignerType.Trezor].includes(signerType)) {
       const signerName = capitalize(signerType)
-      return this.resError(`${signerName} only supports eth_signTypedData_v4+`, payload, res)
+      return resError(`${signerName} only supports eth_signTypedData_v4+`, payload, res)
     }
 
     const handlerId = this.addRequestHandler(res)
@@ -778,12 +645,12 @@ export class Provider extends EventEmitter {
 
       return res({ id: payload.id, jsonrpc: '2.0', result: undefined })
     } catch (e) {
-      return this.resError(e as EVMError, payload, res)
+      return resError(e as EVMError, payload, res)
     }
   }
 
   addEthereumChain (payload: RPCRequestPayload, res: RPCRequestCallback) {
-    if (!payload.params[0]) return this.resError('addChain request missing params', payload, res)
+    if (!payload.params[0]) return resError('addChain request missing params', payload, res)
 
     const type = 'ethereum'
     const { 
@@ -795,9 +662,9 @@ export class Provider extends EventEmitter {
       iconUrls = [] 
     } = payload.params[0]
 
-    if (!chainId) return this.resError('addChain request missing chainId', payload, res)
-    if (!chainName) return this.resError('addChain request missing chainName', payload, res)
-    if (!nativeCurrency) return this.resError('addChain request missing nativeCurrency', payload, res)
+    if (!chainId) return resError('addChain request missing chainId', payload, res)
+    if (!chainName) return resError('addChain request missing chainName', payload, res)
+    if (!nativeCurrency) return resError('addChain request missing nativeCurrency', payload, res)
 
     const handlerId = this.addRequestHandler(res)
 
@@ -830,28 +697,11 @@ export class Provider extends EventEmitter {
     }
   }
 
-  getPermissions (payload: JSONRPCRequestPayload, res: RPCRequestCallback) {
-    const now = new Date().getTime()
-    const toPermission = permission.bind(null, now)
-    const allowedOperations = protectedMethods.map(toPermission)
-
-    res({ id: payload.id, jsonrpc: '2.0', result: allowedOperations })
-  }
-
-  requestPermissions (payload: JSONRPCRequestPayload, res: RPCRequestCallback) {
-    // we already require the user to grant permission to call this method so
-    // we just need to return permission objects for the requested operations
-    const now = new Date().getTime()
-    const requestedOperations = (payload.params || []).map(param => permission(now, Object.keys(param)[0]))
-
-    res({ id: payload.id, jsonrpc: '2.0', result: requestedOperations })
-  }
-
   private addCustomToken (payload: RPCRequestPayload, cb: RPCRequestCallback, targetChain: Chain) {
     const { type, options: tokenData } = (payload.params || {}) as any
 
     if ((type || '').toLowerCase() !== 'erc20') {
-      return this.resError('only ERC-20 tokens are supported', payload, cb)
+      return resError('only ERC-20 tokens are supported', payload, cb)
     }
 
     this.getChainId(payload, resp => {
@@ -862,7 +712,7 @@ export class Provider extends EventEmitter {
       const decimals = parseInt(tokenData.decimals || '1')
 
       if (!address) {
-        return this.resError('tokens must define an address', payload, cb)
+        return resError('tokens must define an address', payload, cb)
       }
 
       const res = () => {
@@ -903,18 +753,6 @@ export class Provider extends EventEmitter {
     }, targetChain)
   }
 
-  private getAssets (payload: RPC.GetAssets.Request, cb: RPCCallback<RPC.GetAssets.Response>) {
-    const currentAccount = accounts.current()
-    if (!currentAccount) return this.resError('no account selected', payload, cb)
-
-    if (isScanning(currentAccount.id)) return this.resError({ message: 'assets not known for account', code: 5901 }, payload, cb)
-
-    const { nativeCurrency, erc20 } = loadAssets(currentAccount.id)
-    const { id, jsonrpc } = payload
-
-    cb({ id, jsonrpc, result: { nativeCurrency, erc20 }})
-  }
-
   private parseTargetChain (payload: RPCRequestPayload) {
     const target: Chain = { type: 'ethereum', id: 0 }
 
@@ -947,20 +785,30 @@ export class Provider extends EventEmitter {
 
   send (payload: RPCRequestPayload, res: RPCRequestCallback = () => {}) {
     const method = payload.method || ''
-
     const targetChain = this.parseTargetChain(payload)
 
     if (!targetChain.id) {
       log.warn('received request with unknown chain', JSON.stringify(payload))
-      return this.resError({ message: `unknown chain: ${payload.chainId}`, code: 4901 }, payload, res)
+      return resError({ message: `unknown chain: ${payload.chainId}`, code: 4901 }, payload, res)
     }
 
-    if (method === 'eth_coinbase') return this.getCoinbase(payload, res)
-    if (method === 'eth_accounts') return this.getAccounts(payload, res)
-    if (method === 'eth_requestAccounts') return this.getAccounts(payload, res)
+    function getAccounts (payload: JSONRPCRequestPayload, res: RPCRequestCallback) {
+      res({ id: payload.id, jsonrpc: payload.jsonrpc, result: accounts.getSelectedAddresses().map(a => a.toLowerCase()) })
+    }
+
+    function getCoinbase (payload: RPCRequestPayload, res: RPCRequestCallback) {
+      accounts.getAccounts((err, accounts) => {
+        if (err) return resError(`signTransaction Error: ${JSON.stringify(err)}`, payload, res)
+        res({ id: payload.id, jsonrpc: payload.jsonrpc, result: (accounts || [])[0] })
+      })
+    }
+
+    if (method === 'eth_coinbase') return getCoinbase(payload, res)
+    if (method === 'eth_accounts') return getAccounts(payload, res)
+    if (method === 'eth_requestAccounts') return getAccounts(payload, res)
     if (method === 'eth_sendTransaction') return this.sendTransaction(payload as RPC.SendTransaction.Request, res)
     if (method === 'eth_getTransactionByHash') return this.getTransactionByHash(payload, res, targetChain)
-    if (method === 'personal_ecRecover') return this.ecRecover(payload, res)
+    if (method === 'personal_ecRecover') return ecRecover(payload, res)
     if (method === 'web3_clientVersion') return this.clientVersion(payload, res)
     if (method === 'eth_subscribe' && payload.params[0] in this.subscriptions) {
       return this.subscribe(payload as RPC.Subscribe.Request, res)
@@ -977,11 +825,11 @@ export class Provider extends EventEmitter {
     
     if (method === 'wallet_addEthereumChain') return this.addEthereumChain(payload, res)
     if (method === 'wallet_switchEthereumChain') return this.switchEthereumChain(payload, res)
-    if (method === 'wallet_getPermissions') return this.getPermissions(payload, res)
-    if (method === 'wallet_requestPermissions') return this.requestPermissions(payload, res)
+    if (method === 'wallet_getPermissions') return getPermissions(payload, res)
+    if (method === 'wallet_requestPermissions') return requestPermissions(payload, res)
     if (method === 'wallet_watchAsset') return this.addCustomToken(payload, res, targetChain)
-    if (method === 'wallet_getChains') return this.getChains(payload, res)
-    if (method === 'wallet_getAssets') return this.getAssets(payload as RPC.GetAssets.Request, res as RPCCallback<RPC.GetAssets.Response>)
+    if (method === 'wallet_getChains') return getChains(payload, res)
+    if (method === 'wallet_getAssets') return getAssets(payload as RPC.GetAssets.Request, accounts.current(), res as RPCCallback<RPC.GetAssets.Response>)
 
     // Connection dependent methods need to pass targetChain
     if (method === 'net_version') return this.getNetVersion(payload, res, targetChain)
@@ -1000,16 +848,6 @@ export class Provider extends EventEmitter {
 }
 
 const provider = new Provider()
-
-function getActiveChains () {
-  const chains: Record<string, Network> = store('main.networks.ethereum') || {}
-
-  return Object.values(chains)
-    .filter(chain => chain.on)
-    .map(chain => chain.id)
-    .sort((a, b) => a - b)
-}
-
 let origins: Origins = store('main.origins')
 let availableChains = getActiveChains()
 
@@ -1053,4 +891,4 @@ store.observer(() => {
   }
 }, 'provider:account')
 
-export default provider
+export { provider }
